refactor(Popover): Use `partials` pattern
