+++ packages/react-component-library/src/components/Popover/partials/StyledFloatingBox.tsx
@@ -1,0 +1,18 @@
+import styled, { css } from 'styled-components'
+
+import { FloatingBox } from '../../../primitives'
+
+interface StyledFloatingBoxProps {
+  $isVisible: boolean
+}
+
+export const StyledFloatingBox = styled(FloatingBox)<StyledFloatingBoxProps>`
+  opacity: 0;
+  transition: linear opacity 0.3s;
+
+  ${({ $isVisible }) =>
+    $isVisible &&
+    css`
+      opacity: 1;
+    `}
+`
